requirements.txt
python-telegram-bot==13.7
python-dotenv==0.19.2
Flask==2.0.1
Werkzeug==2.0.3
urllib3==1.26.5

python-dotenv
pyTelegramBotAPI
pytest


